HOMS-302 Turn Messenger class to importable module
