+++ hbw/app/assets/stylesheets/hbw.sass
@@ -4,6 +4,8 @@
 @import url("https://use.fontawesome.com/releases/v5.7.1/css/regular.css")
 @import url("https://use.fontawesome.com/releases/v5.7.1/css/fontawesome.css")
 
+@import 'jquery.growl'
+
 .hbw-styles
   @import 'hbw/variables'
